feat: add ImageCropper to Mantis
