+++ SwiftUIExample/MantisSwiftUIExample/ContentView.swift
@@ -14,7 +14,7 @@ struct ContentView: View {
     @State private var showingCropShapeList = false
     @State private var cropShapeType: Mantis.CropShapeType = .rect
     @State private var presetFixedRatioType: Mantis.PresetFixedRatioType = .canUseMultiplePresetFixedRatio()
-    @State private var cropperType: MantisImageCropperType = .normal
+    @State private var cropperType: ImageCropper.CropperType = .normal
     @State private var contentHeight: CGFloat = 0
     
     @State private var showImagePicker = false
