--- SwiftUIExample/MantisSwiftUIExample/ImageCropper.swift
@@ -1,58 +1,0 @@
-//
-//  ImageCropper.swift
-//  MantisSwiftUIExample
-//
-//  Created by Yingtao Guo on 2/16/23.
-//
-
-import Mantis
-import SwiftUI
-
-enum MantisImageCropperType {
-    case normal
-    case noRotaionDial
-    case noAttachedToolbar
-}
-
-struct ImageCropper: UIViewControllerRepresentable {
-    let config: Mantis.Config
-    
-    @Binding var image: UIImage?
-    @Binding var transformation: Transformation?
-    @Binding var cropInfo: CropInfo?
-    
-    @Environment(\.presentationMode) var presentationMode
-    
-    class Coordinator: CropViewControllerDelegate {
-        var parent: ImageCropper
-        
-        init(_ parent: ImageCropper) {
-            self.parent = parent
-        }
-        
-        func cropViewControllerDidCrop(_ cropViewController: Mantis.CropViewController, cropped: UIImage, transformation: Transformation, cropInfo: CropInfo) {
-            parent.image = cropped
-            parent.transformation = transformation
-            parent.presentationMode.wrappedValue.dismiss()
-        }
-        
-        func cropViewControllerDidCancel(_ cropViewController: Mantis.CropViewController, original: UIImage) {
-            parent.presentationMode.wrappedValue.dismiss()
-        }
-    }
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator(self)
-    }
-    
-    func makeUIViewController(context: Context) -> UIViewController {
-        let cropViewController = Mantis.cropViewController(image: image!,
-                                                           config: config)
-        cropViewController.delegate = context.coordinator
-        return cropViewController
-    }
-    
-    func updateUIViewController(_ uiViewController: UIViewController, context: Context) {
-        
-    }
-}
